+++ test/dance/core_test.clj
@@ -1,8 +1,13 @@
 (ns dance.core-test
   (:require [clojure.test :refer :all]
-            [dance.core :refer [dance depth-dance break-dance backtrack]]))
+            [dance.core :refer :all]))
 
 (defrecord DanceTestRecord [a b c])
+
+(defmacro printed-lines [& body]
+  `(-> (with-out-str ~@body)
+       (clojure.string/split  #"\n")
+       (->> (mapv #(clojure.string/replace % #"\s+$" "")))))
 
 (deftest test-simple-dance
   (is (= {:a 2 :b 3 :c 4}
@@ -26,6 +31,81 @@
                 :context        {:cnt 0}
                 :return-context true))))
 
+(deftest test-scoped-context
+  (is (= ["- Walking        : (1 [:a] 3)"
+          "  Context        : {:x 0, :depth 0}"
+          "  - Walking        : 1"
+          "    Context        : {:x 0, :depth 1}"
+          "  - Walking        : [:a]"
+          "    Context        : {:x 123, :depth 1, :y 456}"
+          "    - Walking        : :a"
+          "      Context        : {:x 123, :depth 2, :y 456}"
+          "  - Walking        : 3"
+          "    Context        : {:x 0, :depth 1}"]
+         (printed-lines
+           (dance '(1 [:a] 3)
+                  :context {:x 0}
+                  :scoped [:x :y]
+                  :debug true
+                  :pre? (fn [form ctx]
+                          (if (vector? form)
+                            [form (assoc ctx :x 123 :y 456)]
+                            [form ctx]))))))
+  (is (= ["- Walking        : (1 [:a] 3)"
+          "  Context        : {:x 0, :depth 0}"
+          "  - Walking        : 1"
+          "    Context        : {:x 0, :depth 1, :index 0}"
+          "  - Walking        : [:a]"
+          "    Context        : {:x 123, :depth 1, :index 1, :y 456}"
+          "    - Walking        : :a"
+          "      Context        : {:x 123, :depth 2, :index 0, :y 456}"
+          "  - Walking        : 3"
+          "    Context        : {:x 0, :depth 1, :index 2}"]
+         (printed-lines
+           (dance '(1 [:a] 3)
+                  :step step-indexed ;; The only difference with code from above
+                  :context {:x 0}
+                  :scoped [:x :y]
+                  :debug true
+                  :pre? (fn [form ctx]
+                          (if (vector? form)
+                            [form (assoc ctx :x 123 :y 456)]
+                            [form ctx])))))))
+
+(deftest test-with-context-dance
+  (is (= ["- Walking        : [1 #:dance.core{:with-context {:form [:a], :ctx {:x 123, :y 456}}} 3]"
+          "  Context        : {:depth 0}"
+          "  - Walking        : 1"
+          "    Context        : {:depth 1}"
+          "  - Walking        : [:a]"
+          "    Context        : {:depth 1, :x 123, :y 456}"
+          "    - Walking        : :a"
+          "      Context        : {:depth 2, :x 123, :y 456}"
+          "  - Walking        : 3"
+          "    Context        : {:depth 1, :y 456}"]
+         (printed-lines
+           (dance [1 (with-context [:a] {:x 123 :y 456}) 3]
+                  with-context-dance
+                  :scoped [:x]
+                  ; :debug-context true
+                  ))))
+  (is (= ["- Walking        : [1 #:dance.core{:with-context {:form [:a], :ctx {:x 123, :y 456, :scoped [:y]}}} 3]"
+          "  Context        : {:x 0, :depth 0}"
+          "  - Walking        : 1"
+          "    Context        : {:x 0, :depth 1}"
+          "  - Walking        : [:a]"
+          "    Context        : {:x 123, :depth 1, :y 456, :scoped [:y]}"
+          "    - Walking        : :a"
+          "      Context        : {:x 123, :depth 2, :y 456, :scoped [:y]}"
+          "  - Walking        : 3"
+          "    Context        : {:x 0, :depth 1, :scoped [:y]}"]
+         (printed-lines
+           (dance [1 (with-context [:a] {:x 123 :y 456 :scoped [:y]}) 3]
+                  with-context-dance
+                  :context {:x 0}
+                  :scoped [:x]
+                  :debug-context true)))))
+
 (deftest test-dance-on-record
   (let [result (dance (DanceTestRecord. 1 2 [3 4 5])
                       :post? number?
@@ -33,16 +113,16 @@
     (is (instance? DanceTestRecord result))
     (is (= result (DanceTestRecord. 2 3 [4 5 6])))))
 
-(deftest test-break-dance
+(deftest test-break-dance!
   (is (= [3 4 :needle [5 6]]
          (dance [1 2 [3 4 :needle [5 6]]]
                 :pre? coll?
                 :pre #(or (when (.contains % :needle)
-                             (break-dance %))
+                             (break-dance! %))
                            %)))))
 
 (deftest test-backtrack
-  (testing "no params"
+  #_(testing "no params"
     (is (= [2 [2] [[1] [[[1]]]]]
            (dance [1 [1] [[1] [[[1]]]]]
                   depth-dance
@@ -52,7 +132,7 @@
                               [x ctx]))
                   :post? number?
                   :post inc))))
-  (testing "one param (form)"
+  #_(testing "one param (form)"
     (is (= [2 [101] [101 101]]
            (dance [1 [1] [[1] [[[1]]]]]
                   depth-dance
@@ -62,7 +142,7 @@
                               [x ctx]))
                   :post? number?
                   :post inc))))
-  (testing "two params (form ctx)"
+  #_(testing "two params (form ctx)"
     (is (= [[2 [101] [101 101]] {:abc :xyz}]
            (dance [1 [1] [[1] [[[1]]]]]
                   depth-dance
@@ -73,11 +153,11 @@
                   :post? number?
                   :post inc
                   :return-context true))))
-  (comment "later"
-    (testing "anticipated backtracking"
-      (pprint (dance [1 2 (backtrack [3 4])]
-                     depth-dance
-                     :debug true
-                     :post? number?
-                     :post inc
-                     :return-context true)))))
+  #_(testing "anticipated backtracking"
+    (println "--->" (backtrack [3 4]))
+    #_(pprint (dance [1 2 (backtrack [3 4])]
+                   depth-dance
+                   :debug true
+                   :post? number?
+                   :post inc
+                   :return-context true))))