+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [Unreleased]
+### Added
+- `:unscoped` option. Like `:scoped`, applies to both dances and their context.
+
+### Changed
+- Renamed `break-dance` to `break-dance!` (could confusedly thought to be a
+  dance).
+
 ## [0.1.1] 2018-05-01
-## Fixed
+### Fixed
 - Reworked dependencies on shuriken.
 
 ## [0.1.0] 2018-05-01